refactor: Logger class moved to custom_logger.__init__
